requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
email-validator==2.1.0
sqlalchemy==2.0.36
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart==0.0.9
google-generativeai==0.8.3




